#262 Split-off ResourceManagerClassLoader
