+++ piranha/src/main/java/com/manorrock/piranha/DefaultWebApplicationClassLoader.java
@@ -27,34 +27,17 @@
  */
 package com.manorrock.piranha;
 
-import com.manorrock.piranha.api.WebApplicationClassLoader;
-import com.manorrock.piranha.api.ResourceManager;
-import java.io.BufferedInputStream;
-import java.io.ByteArrayOutputStream;
 import java.io.File;
-import java.io.IOException;
-import java.net.MalformedURLException;
-import java.net.URL;
-import java.util.Collections;
-import java.util.Enumeration;
-import java.util.concurrent.ConcurrentHashMap;
+
+import com.manorrock.piranha.api.ResourceManager;
+import com.manorrock.piranha.api.WebApplicationClassLoader;
 
 /**
  * The default WebApplicationClassLoader.
  *
  * @author Manfred Riem (mriem@manorrock.com)
  */
-public class DefaultWebApplicationClassLoader extends ClassLoader implements WebApplicationClassLoader {
-
-    /**
-     * Stores the resource manager.
-     */
-    private ResourceManager resourceManager;
-
-    /**
-     * Stores the loaded classes.
-     */
-    private final ConcurrentHashMap<String, Class> classes = new ConcurrentHashMap<>();
+public class DefaultWebApplicationClassLoader extends DefaultResourceManagerClassLoader implements WebApplicationClassLoader {
 
     /**
      * Constructor.
@@ -68,11 +51,12 @@ public class DefaultWebApplicationClassLoader extends ClassLoader implements Web
      * @param baseDirectory the base directory.
      */
     public DefaultWebApplicationClassLoader(File baseDirectory) {
-        resourceManager = new DefaultResourceManager();
+        ResourceManager resourceManager = new DefaultResourceManager();
         File classesDirectory = new File(baseDirectory, "WEB-INF/classes");
         if (classesDirectory.exists()) {
             resourceManager.addResource(new DefaultDirectoryResource(classesDirectory));
         }
+        
         File libDirectory = new File(baseDirectory, "WEB-INF/lib");
         if (libDirectory.exists()) {
             File[] jarFiles = libDirectory.listFiles();
@@ -82,95 +66,8 @@ public class DefaultWebApplicationClassLoader extends ClassLoader implements Web
                 }
             }
         }
+        
+        setResourceManager(resourceManager);
     }
-
-    /**
-     * Constructor.
-     *
-     * @param resourceManager the resource manager.
-     */
-    public DefaultWebApplicationClassLoader(ResourceManager resourceManager) {
-        this.resourceManager = resourceManager;
-    }
-
-    /**
-     * Load the class.
-     *
-     * @param name the name.
-     * @param resolve the resolve flag.
-     * @return the class.
-     * @throws ClassNotFoundException when the class cannot be found.
-     */
-    @Override
-    protected Class<?> loadClass(String name, boolean resolve) throws ClassNotFoundException {
-        Class<?> result;
-        try {
-            result = getSystemClassLoader().loadClass(name);
-        } catch (ClassNotFoundException cnfe) {
-            result = null;
-        }
-        if (result == null) {
-            if (classes.containsKey(name)) {
-                result = classes.get(name);
-            } else {
-                try {
-                    String normalizedName = name.replaceAll("\\.", "/") + ".class";
-                    BufferedInputStream inputStream = new BufferedInputStream(resourceManager.getResourceAsStream(normalizedName));
-                    ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
-                    int read = inputStream.read();
-                    while (read != -1) {
-                        outputStream.write((byte) read);
-                        read = inputStream.read();
-                    }
-                    byte[] bytes = outputStream.toByteArray();
-                    result = defineClass(name, bytes, 0, bytes.length);
-                    if (resolve) {
-                        resolveClass(result);
-                    }
-                    classes.put(name, result);
-                } catch (Throwable throwable) {
-                    throw new ClassNotFoundException("Unable to load class: " + name, throwable);
-                }
-            }
-        }
-        return result;
-    }
-
-    /**
-     * Find the resource.
-     *
-     * @param name the name.
-     * @return the resource, or null if not found.
-     */
-    @Override
-    protected URL findResource(String name) {
-        URL result = null;
-        try {
-            result = resourceManager.getResource(name);
-        } catch (MalformedURLException mue) {
-        }
-        return result;
-    }
-
-    /**
-     * Find the resources.
-     *
-     * @param name the name of the resource.
-     * @return the enumeration of the resource urls.
-     * @throws IOException when an I/O error occurs.
-     */
-    @Override
-    protected Enumeration<URL> findResources(String name) throws IOException  {
-        return Collections.enumeration(resourceManager.getResources(name));
-    }
-
-    /**
-     * Set the resource manager.
-     *
-     * @param resourceManager the resource manager.
-     */
-    @Override
-    public void setResourceManager(ResourceManager resourceManager) {
-        this.resourceManager = resourceManager;
-    }
+    
 }
